msg/async/rdma: record remote QP number
When remote QP is destroyed, QP will be disconnected. The local QP
is transitioned into error state. Then some asynchronous event or
completion event could be triggered. It's better to out the remote
QP number in this case to check whether the event is triggered because
remote QP is destroyed.
Add get_remote_qpn to get qpn through RDMAConnectedSocketImpl object.

Signed-off-by: Changcheng Liu <changcheng.liu@aliyun.com>